Refactor validation into own module.
